requirements.txt
# Core AI / ML
transformers>=4.40.0
torch>=2.2.0          # CPU-only: torch>=2.2.0+cpu (see README)
sentencepiece>=0.1.99 # required by some HF tokenizers

# Computer vision
opencv-python>=4.9.0

# Document parsing
pdfplumber>=0.10.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2

# Scheduling
schedule>=1.2.1

# Config & utilities
PyYAML>=6.0.1
colorlog>=6.8.2

# Testing
pytest>=8.1.0
pytest-cov>=5.0.0
